change double to bigdecimal

+++ src/main/java/com/my/stock/rdb/entity/Stock.java
@@ -5,6 +5,8 @@ import com.my.stock.base.entity.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @EqualsAndHashCode(callSuper = false)
 @Builder
 @Entity
@@ -20,9 +22,9 @@ public class Stock extends BaseTimeEntity {
 	@Column(nullable = false)
 	private String symbol;
 	@Column(nullable = false)
-	private double quantity;
+	private BigDecimal quantity;
 	@Column(nullable = false)
-	private double price;
+	private BigDecimal price;
 
 	@ManyToOne
 	private BankAccount bankAccount;
